requirements.txt
numpy==1.19.3
Pillow==8.3.1
